requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.8
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.6
MarkupPy==1.14
multidict==6.0.4
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.9.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
